refactror(#18): compound 패턴 적용 및 큰 organism을 molecule/organism 컴포넌트들로 분리
